Project refactor
