+++ src/components/Pages/Upload/UploadSplitsPage.tsx
@@ -1,6 +1,6 @@
 import "./UploadSplitsPage.css";
 import React, { useState } from "react";
-import Splits from "../classes/Splits";
+import Splits from "../../../classes/Splits";
 
 interface UploadSplitsProps {
   setFile: React.Dispatch<React.SetStateAction<File | null>>;
